requirements.txt
# Core Flask components
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3

# Database and ORM
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
alembic==1.12.0

# Authentication and security
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
WTForms==3.0.1
email-validator==2.1.0

# Environment and configuration
python-dotenv==1.0.0
environs==9.5.0

# Development tools
pytest==7.4.2
pytest-flask==1.2.0
flake8==6.1.0
black==23.9.1

# Deployment tools
gunicorn==21.2.0
